Added some bulletproofing to memory allocation in the LDAP cache code. PR: Obtained from: Submitted by: Reviewed by:
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@90789 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/ldap/util_ldap_cache.h
@@ -193,7 +193,7 @@ void util_ldap_dn_compare_node_free(void *n);
 /* util_ldap_cache_mgr.c */
 
 void util_ald_free(const void *ptr);
-void *util_ald_alloc(int size);
+void *util_ald_alloc(unsigned long size);
 const char *util_ald_strdup(const char *s);
 unsigned long util_ald_hash_string(int nstr, ...);
 void util_ald_cache_purge(util_ald_cache_t *cache);
